requirements.txt
django
python-dotenv
djangorestframework
drf-yasg
langchain
langchain-community
openai
faiss-cpu
chromadb
tiktoken
django-cors-headers
langgraph
langchain-openai